requirements.txt
aiogram
python-dotenv
tabulate
motor
pymongo
